refactor: move sidebar components to feature folder
