+++ packages/api-reference/src/features/sidebar/components/SidebarElement.vue
