+++ packages/api-reference/src/features/sidebar/index.ts
@@ -1,2 +1,4 @@
+export { default as Sidebar } from './components/Sidebar.vue'
+
 export { useSidebar, SIDEBAR_SYMBOL } from './hooks/useSidebar'
 export { createSidebar, type SidebarOptions } from './helpers/create-sidebar'
